chore(memory): cleared all the dynamic allocation, no more memory leaks upon exit
