+++ src/init/init.c
@@ -26,6 +26,7 @@ void    init(t_game *game)
     game->s_path[2] = ft_strdup("./textures/Sprite3.xpm");
     game->first_mouse = 1;
     game->t_path[T_COUNT] = NULL;
+    game->s_path[FRAMES] = NULL;
     game->p.p_x = 10.5;
 	game->p.p_y = 6.5;
     game->sprite.x = 10.5;
