+++ src/utils/cleanup.c
@@ -25,4 +25,6 @@ void    cleanup(t_game *game)
         free_2d((void **)game->t_path);
     if (game->doors)
         free_2d((void **)game->doors);
+    if (game->s_path)
+        free_2d((void **)game->s_path);
 }
